refactor: retire legacy worker mode and qualify shared transport
