+++ scripts/pr-risk-classifier.mjs
@@ -8,8 +8,7 @@ const retainedExecutionWebPaths = new Set([
   "web/authoring-client.js",
   "web/python-worker.source.js",
   "web/semantic-engine-endpoint.js",
-  "web/execution-engine-worker.js",
-  "scripts/retained-execution-worker-smoke.mjs",
+  "scripts/execution-worker-smoke.mjs",
   "web/authoring-execution-client.js",
   "web/authoring-render-worker.js",
   "web/execution-canvas.js",
